Extract initialization logic into its own module
